chore: refactor Leaderboard to server component, add util function to fetch data
