+++ app/leaderboard/page.tsx
@@ -1,4 +1,5 @@
 import Leaderboard from "./Leaderboard";
+("use client");
 
 export default function Home(): JSX.Element {
   return (
